added delete ooption & elevation now in float

+++ src-tauri/src/io.rs
@@ -42,14 +42,14 @@ pub fn read_track_analysis(ulid: &String) -> Result<TrackAnalysis, io::Error> {
     Ok(ta)
 }
 
-pub fn read_elevation(ulid: &str) -> Result<Vec<(f64, i32)>, io::Error> {
+pub fn read_elevation(ulid: &str) -> Result<Vec<(f64, f64)>, io::Error> {
     let path = paths::track_elevation(ulid);
     let json_string = fs::read_to_string(path)?;
-    let elevation: Vec<(f64, i32)> = serde_json::from_str(&json_string.as_str())?;
+    let elevation: Vec<(f64, f64)> = serde_json::from_str(&json_string.as_str())?;
     Ok(elevation)
 }
 
-pub fn write_elevation(elevation: Vec<(f64, i32)>, ulid: &str) -> Result<(), io::Error> {
+pub fn write_elevation(elevation: Vec<(f64, f64)>, ulid: &str) -> Result<(), io::Error> {
     let path = paths::track_elevation(ulid);
     write_file(path, serde_json::to_string(&elevation)?)?;
     Ok(())
